feat(hose): add stock for product
