+++ APP.Eds/APP.Eds/Components/PopUp/AddDispenser.xaml
@@ -13,6 +13,7 @@
         <ResourceDictionary>
             <converters:DecimalFormatConverter x:Key="DecimalFormat" />
             <converters:CurrencyFormatConverter x:Key="CurrencyFormat" />
+            <converters:IsNotNullConverter x:Key="IsNotNullConverter" />
         </ResourceDictionary>
     </mct:Popup.Resources>
 
@@ -24,7 +25,7 @@
            BackgroundColor="White"
            WidthRequest="350" 
            MaximumWidthRequest="380"
-           MaximumHeightRequest="580"
+           MaximumHeightRequest="620"
            VerticalOptions="Center" 
            HorizontalOptions="Center"
            Margin="15,30">
@@ -80,6 +81,40 @@
                                     BackgroundColor="Transparent" FontSize="12"
                                     HorizontalTextAlignment="Center"/>
                         </Frame>
+                    </StackLayout>
+                </Frame>
+
+                <!-- Stock del Producto (solo lectura) -->
+                <Frame x:Name="StockInfoFrame" BackgroundColor="#F3E5F5" CornerRadius="10" HasShadow="True"
+                       BorderColor="#E1BEE7" Padding="10" IsVisible="{Binding SelectedHose, Converter={StaticResource IsNotNullConverter}}">
+                    <StackLayout Spacing="6">
+                        <StackLayout Orientation="Horizontal" Spacing="6">
+                            <Frame BackgroundColor="#E1BEE7" CornerRadius="8" HasShadow="False"
+                                   Padding="0" WidthRequest="16" HeightRequest="16" VerticalOptions="Center">
+                                <Label Text="📦" FontSize="8" HorizontalOptions="Center" VerticalOptions="Center"/>
+                            </Frame>
+                            <Label Text="Stock del Producto" FontSize="13" FontAttributes="Bold"
+                                   TextColor="#2D3436" VerticalOptions="Center"/>
+                        </StackLayout>
+
+                        <Grid RowDefinitions="*,*" ColumnDefinitions="*,15,*" RowSpacing="4">
+                            <Label Text="Producto" VerticalOptions="Center"
+                                   HorizontalOptions="End" FontSize="11" Grid.Row="0" Grid.Column="0"/>
+                            <Label Text="{Binding SelectedHose.ProductEntity.Name}" 
+                                   VerticalOptions="Center" FontSize="11" FontAttributes="Bold"
+                                   TextColor="#6A1B9A" Grid.Row="0" Grid.Column="2"/>
+
+                            <Label Text="Stock disponible" VerticalOptions="Center"
+                                   HorizontalOptions="End" FontSize="11" FontAttributes="Bold"
+                                   Grid.Row="1" Grid.Column="0"/>
+                            <StackLayout Orientation="Horizontal" Spacing="2" Grid.Row="1" Grid.Column="2">
+                                <Label Text="{Binding SelectedHose.ProductEntity.Stock, StringFormat='{0:N0}'}" 
+                                       VerticalOptions="Center" FontSize="11" FontAttributes="Bold"
+                                       TextColor="#6A1B9A"/>
+                                <Label Text="G" VerticalOptions="Center" FontSize="9" 
+                                       TextColor="#6A1B9A" FontAttributes="Bold"/>
+                            </StackLayout>
+                        </Grid>
                     </StackLayout>
                 </Frame>
