+++ APP.Eds/APP.Eds/Converters/ValidationConverters.cs
@@ -3,19 +3,6 @@ using System.Collections;
 
 namespace APP.Eds.Converters
 {
-    public class IsNotNullConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            return value != null;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
     public class IsNotNullOrEmptyConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
